chore(server): apply clippy autofixes (191->57 warnings)
Mechanical clippy --fix run for Rust 1.94 / edition 2024:
- Collapsible if-let into let-chains
- Replace manual range-contains, is-multiple-of, div-ceil, abs-diff
  with std equivalents
- Drop redundant closures and same-type casts
- Switch side-effect map() to inspect()

Also remove dead loop {} wrapper around tidal_auth::poll_for_token
in TIDAL device-flow spawn -- the inner function already polls.

+++ noor-server/src/services/musicbrainz.rs
@@ -515,7 +515,7 @@ pub fn import_portable_snapshot(conn: &Connection) -> Result<PortableSnapshotImp
             checked_skipped += 1;
             continue;
         };
-        checked_inserted += insert_checked.execute(params![track_id])? as usize;
+        checked_inserted += insert_checked.execute(params![track_id])?;
     }
 
     let mut genre_inserted = 0usize;
@@ -552,7 +552,7 @@ pub fn import_portable_snapshot(conn: &Connection) -> Result<PortableSnapshotImp
             genre_skipped += 1;
             continue;
         };
-        genre_inserted += insert_genre.execute(params![track_id, genre_id, confidence])? as usize;
+        genre_inserted += insert_genre.execute(params![track_id, genre_id, confidence])?;
     }
     drop(insert_genre);
     drop(insert_checked);
